requirements.txt
asgiref==3.8.1
Django==4.2.13
django-phonenumber-field==8.0.0
phonenumberslite==8.13.40
pillow==10.4.0
sqlparse==0.5.1
